Refactor generate_schema to tsdat.cli
