+++ test/test_cli.py
@@ -2,7 +2,7 @@ import tempfile
 
 from typer.testing import CliRunner
 
-from tsdat.main import app
+from tsdat.cli import app
 
 runner = CliRunner()
 
@@ -12,7 +12,7 @@ def test_schema_generation():
     with tempfile.TemporaryDirectory() as tmp_dir:
         result = runner.invoke(app, ["generate-schema", "--dir", tmp_dir])
         assert result.exit_code == 0
-        assert f"tsdat dataset standards" in result.stdout
+        assert "tsdat dataset standards" in result.stdout
 
     # acdd
     with tempfile.TemporaryDirectory() as tmp_dir:
@@ -20,7 +20,7 @@ def test_schema_generation():
             app, ["generate-schema", "--dir", tmp_dir, "--standards", "acdd"]
         )
         assert result.exit_code == 0
-        assert f"acdd dataset standards" in result.stdout
+        assert "acdd dataset standards" in result.stdout
 
     # ioos
     with tempfile.TemporaryDirectory() as tmp_dir:
@@ -28,4 +28,4 @@ def test_schema_generation():
             app, ["generate-schema", "--dir", tmp_dir, "--standards", "ioos"]
         )
         assert result.exit_code == 0
-        assert f"ioos dataset standards" in result.stdout
+        assert "ioos dataset standards" in result.stdout
