+++ tsdat/__main__.py
@@ -1,3 +1,3 @@
-from .main import app  # pragma: no cover
+from .cli import app  # pragma: no cover
 
 app(prog_name="tsdat")  # pragma: no cover
