+++ tsdat/cli/__init__.py
@@ -1,1 +1,2 @@
-from .generate_schema import generate_schema as generate_schema
+from .cli import app as app
+from .generate_schema.generate_schema import generate_schema as generate_schema
